feat: make `simapp.MakeTestEncodingConfig` private
